requirements.txt

aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
config==0.5.0.post0
dateparser==1.0.0
Django==3.2.6
django-bootstrap4==3.0.1
idna==3.2
multidict==5.1.0
numpy==1.21.2
pandas==1.3.2
pandas-ta==0.3.14b0
pprintpp==0.4.0
python-binance==1.0.12
python-dateutil==2.8.2
python-math==0.0.1
pytz==2021.1
regex==2021.8.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
tzdata==2021.1
tzlocal==3.0
ujson==4.1.0
urllib3==1.26.6
websocket-client==1.2.1
websockets==9.1
yarl==1.6.3
